--- batch/src/main/java/com/zipsoon/batch/dto/ResponseDto.java
@@ -1,4 +1,0 @@
-package com.zipsoon.batch.dto;
-
-public class ResponseDto {
-}
