+++ batch/src/main/java/com/zipsoon/batch/estate/mapper/EstateSnapshotMapper.java
