+++ batch/src/main/java/com/zipsoon/batch/estate/repository/EstateSnapshotRepository.java
